refactor: consolidate Google authentication logic and remove duplicate code
